# documentation/openapi.yaml
openapi: 3.0.0
info:
  title: ImageCraft API
  description: API for image processing and storage with user authentication
  version: 1.2.0
servers:
  - url: http://localhost:3001
    description: Development server
paths:
...
  /images/{key}:
    get:
      summary: Get image
      description: Retrieve a specific image
      tags:
        - Images
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
        - in: query
          name: width
          schema:
            type: integer
        - in: query
          name: height
          schema:
            type: integer
        - in: query
          name: fit
          schema:
            type: string
            enum: [cover, contain, fill, inside, outside]
        - in: query
          name: quality
          schema:
            type: integer
      responses:
        '200':
          description: Image retrieved successfully
          content:
            image/webp:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found

    delete:
      summary: Delete image
      description: Delete a specific image
      tags:
        - Images
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Image deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Insufficient permissions or API key users can only delete their own images
        '404':
          description: Image not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...